Cambios y ajustes en algunas clases

+++ src/main/java/service/TarifaService.java
@@ -3,7 +3,7 @@ package service;
 public class TarifaService {
 
     // Tarifa base
-    public static final double TARIFA_POR_KM = 1000.0;
+    public static final int TARIFA_POR_KM = 1000;
 
     // Extras por servicios adicionales
     public static final double EXTRA_SEGURO = 0.15;
@@ -12,14 +12,14 @@ public class TarifaService {
     public static final double EXTRA_PRIORIDAD = 0.20;
 
     // Cálculo por peso
-    public static final double PESO_UMBRAL = 5;
-    public static final double COSTO_POR_KG_EXCEDENTE = 2000;
+    public static final int PESO_UMBRAL = 5;
+    public static final int COSTO_POR_KG_EXCEDENTE = 2000;
 
     // Cálculo por dimensiones
     // Volumen M3
     public static final double VOLUMEN_UMBRAL = 0.1;
-    public static final double COSTO_POR_M3_EXCEDENTE = 5000;
+    public static final int COSTO_POR_M3_EXCEDENTE = 5000;
     // Área M2
     public static final double AREA_UMBRAL = 0.5;
-    public static final double COSTO_POR_M2_EXCEDENTE = 3000.0;
+    public static final int COSTO_POR_M2_EXCEDENTE = 3000;
 }
